requirements.txt
anthropic>=0.40.0
pydantic>=2.0.0
pandas>=2.0.0
python-dotenv>=1.0.0
rich>=13.0.0
typer>=0.12.0
gspread>=6.0.0
google-auth>=2.0.0
